requirements.txt
fastmcp>=2.11.3
requests>=2.32.5
python-dotenv>=1.1.1
git+https://github.com/peterlarnholt/fastmcp-oauth.git
